+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ai-gui/core
 
+## 0.8.0
+
 ## 0.7.0
 
 ### Minor Changes