+++ packages/plugin-citation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-gui/plugin-citation
 
+## 0.8.0
+
+### Patch Changes
+
+- @ai-gui/core@0.8.0
+
 ## 0.7.0
 
 ### Patch Changes